app/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-spectacular
django-cors-headers
python-dotenv
gunicorn
whitenoise
pytest
pytest-django
